% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBatchViewer.R
\docType{package}
\name{MsBatchViewer-package}
\alias{MsBatchViewer}
\alias{MsBatchViewer-package}
\title{MsBatchViewer}
\description{
Shiny tool for charge state resolved total ion current chromatograms
}
\author{
\strong{Maintainer}: Jonas D. Förster \email{jonasfoerster@gmail.com}

}
